// types/banner.ts
export interface Banner {
  largeText2: string;
  id: string;
    buttonText: string;
    product: string;
    image: { url: string };
    description: string;
    smallText: string;
  midText: string;
  largeText: string;
  largetText2: string;
  discount: string;
    saleTime: string;
    updatedAt: string;
    updatedBy: { id: string; name: string };
}